# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/utente": {
      "post": {
        "operationId": "UtenteController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUtenteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtenteDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "User with email already exists"
          }
        },
        "summary": "Create a new user (candidate or interviewer)",
        "tags": [
          "Utenti"
        ]
      },
      "get": {
        "operationId": "UtenteController_findAll",
        "parameters": [
          {
            "name": "exclude",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "id",
                  "nome",
                  "cognome",
                  "email",
                  "ruolo",
                  "intervisteRicevute",
                  "intervisteEffettuate",
                  "selezionatoIn"
                ]
              }
            }
          },
          {
            "name": "fullRelations",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "nome",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "cognome",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "email"
              }
            }
          },
          {
            "name": "ruolo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "intervisteRicevute",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "intervisteEffettuate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "selezionatoIn",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "number",
                "minimum": 0
              }
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseUtenteDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Retrieve users with filters, projection, and pagination",
        "tags": [
          "Utenti"
        ]
      }
    },
    "/utente/{id}": {
      "get": {
        "operationId": "UtenteController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "exclude",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "id",
                  "nome",
                  "cognome",
                  "email",
                  "ruolo",
                  "intervisteRicevute",
                  "intervisteEffettuate",
                  "selezionatoIn"
                ]
              }
            }
          },
          {
            "name": "fullRelations",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtenteDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get a single user by ID",
        "tags": [
          "Utenti"
        ]
      },
      "patch": {
        "operationId": "UtenteController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUtenteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtenteDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Update an existing user",
        "tags": [
          "Utenti"
        ]
      },
      "delete": {
        "operationId": "UtenteController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUtenteDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Delete a user by ID",
        "tags": [
          "Utenti"
        ]
      }
    },
...
  },
  "info": {
    "title": "IntervistaServer API",
    "description": "API specification for IntervistaServer: candidate management, recruitment searches, and conflict-validated interview scheduling.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUtenteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "nome": {
            "type": "string"
          },
          "cognome": {
            "type": "string"
          },
          "ruolo": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "nome",
          "cognome"
        ]
      },
      "ResponseUtenteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "intervisteRicevute": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "intervisteEffettuate": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "selezionatoIn": {
            "type": "object"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "nome": {
            "type": "string"
          },
          "cognome": {
            "type": "string"
          },
          "ruolo": {
            "type": "string"
          }
        },
        "required": [
          "intervisteRicevute",
          "intervisteEffettuate",
          "email",
          "nome",
          "cognome"
        ]
      },
      "UpdateUtenteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "nome": {
            "type": "string"
          },
          "cognome": {
            "type": "string"
          },
          "ruolo": {
            "type": "string"
          }
        }
      },
...
    }
  }
}